+++ JanuszMarcinik.Mvc.Domain/Application/Entities/Questionnaires/Result.cs
@@ -6,18 +6,18 @@ namespace JanuszMarcinik.Mvc.Domain.Application.Entities.Questionnaires
     [Table("Results", Schema = "Questionnaire")]
     public class Result : IApplicationEntity
     {
-        public long ResultId { get; set; }
+        public int ResultId { get; set; }
 
-        public long IntervieweeId { get; set; }
+        public int IntervieweeId { get; set; }
         public virtual Interviewee Interviewee { get; set; }
 
-        public long QuestionnaireId { get; set; }
+        public int QuestionnaireId { get; set; }
         public virtual Questionnaire Questionnaire { get; set; }
 
-        public long QuestionId { get; set; }
+        public int QuestionId { get; set; }
         public virtual Question Question { get; set; }
 
-        public long AnswerId { get; set; }
+        public int AnswerId { get; set; }
         public virtual Answer Answer { get; set; }
     }
 }
