+++ JanuszMarcinik.Mvc.WebUI/Areas/Application/Controllers/SurveyController.cs
@@ -74,7 +74,7 @@ namespace JanuszMarcinik.Mvc.WebUI.Areas.Application.Controllers
         {
             var model = new QuestionnaireViewModel();
             model.SetQuestionnaire(_questionnairesRepository.GetFullModel(questionaireNumber));
-            model.SelectedValues = new List<long>();
+            model.SelectedValues = new List<int>();
             model.QuestionnairesCount = _questionnairesRepository.GetOnlyActives().Count();
 
             return View(model);
